rendering new Student homepage
